pyproject.toml
[tool.poetry]
name = "medical"
version = "0.1.0"
description = ""
authors = ["amirreza <amir.rsf1380@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
django-environ = "^0.11.2"
django-allauth = "^0.63.6"
django-autoslug = "^1.9.9"
django-taggit = "^5.0.1"
psycopg = {extras = ["c"], version = "^3.2.1"}
psycopg-pool = "^3.2.2"
pillow = "^10.4.0"
pillow-avif-plugin = "^1.4.6"
argon2-cffi = "^23.1.0"
django-htmx = "^1.18.0"
daphne = "^4.1.2"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^8.0.0"}
django-jalali = "^7.0.0"

[tool.poetry.group.cms.dependencies]
django-cms = "^4.1.1"
djangocms-text-ckeditor = "^5.1.5"
djangocms-frontend = "^1.3.3"
django-filer = "^3.1.3"
djangocms-alias = "^2.0.1"
djangocms-admin-style = "^3.3.1"
django-sekizai = "^4.1.0"


[tool.poetry.group.dev.dependencies]
whitenoise = {extras = ["brotli"], version = "^6.7.0"}
django-debug-toolbar = "^4.4.6"
black = "^24.4.2"
ruff = "^0.5.2"
mypy = "^1.10.1"
django-stubs = "^5.0.2"
pytest = "^8.2.2"
pytest-django = "^4.8.0"
pytest-sugar = "^1.0.0"
coverage = "^7.6.0"
django-coverage-plugin = "^3.1.0"
pip-tools = "^7.4.1"


[tool.poetry.group.production.dependencies]
gunicorn = "^22.0.0"
boto3 = {extras = ["crt"], version = "^1.34.144"}
django-storages = "^1.14.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4

# Assume Python 3.8
target-version = "py312"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E4", "E7", "E9", "F"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "F401", "F403"]
"**/{migrations}/*" = ["F401"]

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = false

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "django_project.config.test"
python_files = ["test_*.py", "*_test.py", "tests.py"]
